requirements.txt
Flask==3.0.2
Flask-MySQL==1.5.2
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
bcrypt==4.1.2 